Few little things
I took out the text about not replying to emails - because I'm going to set the script to come from a valid email address. I *want* people to be able to reply.

Also, what about moving the CSS out of the PHP script and make it something customisable in the .ini file? (And while I was there, I put a few ifdef() options in to give some default settings.)